+++ core/src/main/python/discover.py
@@ -30,7 +30,7 @@ from wlsdeploy.aliases.wlst_modes import WlstModes
 from wlsdeploy.exception import exception_helper
 from wlsdeploy.logging.platform_logger import PlatformLogger
 from wlsdeploy.util import getcreds
-from wlsdeploy.tool.util.variable_file_helper import VariableFileHelper
+from wlsdeploy.tool.util.variable_injector import VariableFileHelper
 from wlsdeploy.tool.discover import discoverer
 from wlsdeploy.tool.discover.deployments_discoverer import DeploymentsDiscoverer
 from wlsdeploy.tool.discover.domain_info_discoverer import DomainInfoDiscoverer
@@ -372,12 +372,12 @@ def __check_and_customize_model(model, model_context):
 
     default_variable_file = __get_default_variable_file(model_context)
     inserted, variable_model, variable_file_name = VariableFileHelper(model.get_model(), model_context, WebLogicHelper(
-        __logger).get_actual_weblogic_version()).replace_variables_file(variable_file_name=default_variable_file)
+        __logger).get_actual_weblogic_version()).inject_variables_keyword_file(variable_file_name=default_variable_file)
     if default_variable_file:
         default_variable_file = os.path.join(path_utils.get_pathname_from_path(model_context.get_archive_file_name()),
                                              default_variable_file + '.properties')
     inserted, variable_model, variable_file_name = VariableFileHelper(model.get_model(), model_context, WebLogicHelper(
-        __logger).get_actual_weblogic_version()).replace_variables_file(variable_file_name=default_variable_file)
+        __logger).get_actual_weblogic_version()).inject_variables_keyword_file(variable_file_name=default_variable_file)
     if inserted:
         model = Model(variable_model)
     try:
