+++ core/src/main/resources/oracle/weblogic/deploy/messages/wlsdeploy_rb.properties
@@ -1098,7 +1098,7 @@ WLSDPLY-19305=Failed to extract domain library {0} because it does not exist in 
 WLSDPLY-19306=Unable to extract domain library {0} from archive file {1}: {2}
 WLSDPLY-19307=Unable to extract classpath libraries from archive file {0} to domain directory {1}: {2}
 
-# wlsdeploy/tool/util/variable_file_helper.py
+# wlsdeploy/tool/util/variable_injector.py
 WLSDPLY-19400=model variable replacement helper file found at location {0}
 WLSDPLY-19401=Variable replacement using custom list loaded from variable helper file {0}
 WLSDPLY-19402=Unable to read and parse variables helper file {0} : {1}
